requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
psycopg2-binary==2.9.1
    # via -r requirements.in
python-decouple==3.4
    # via -r requirements.in
simple-term-menu==1.2.1
    # via -r requirements.in
